Rewrite "Using submodules" to simplify topic and improve usability
Since the instructions for Codacy Cloud and Self-hosted were
practically the same (except for updating the license) I merged the
procedures, simplifying the page.

I also made a clear distinction between what must be done when
activating submodules for the first time and when adding new
repositories that use submodules.