Fix broken images in older documentation versions
The Markdown in older documentation versions references the pool of
images in docs/images using absolute links.

What this means is that the older documentation versions are actually
using the images published in the "Latest" version.

As I was deleting these images as I updated them and moved them closer
to the Markdown files, the images were disappearing from the older
documentation versions.

To avoid this, I'm bringing back all images to the docs/images folder,
at least until this is fixed in a more definitive way.