clean: Merge "I added a repository" with "Getting started"
The page "I added a repository" is the logical conclusion to the steps
performed on the "Getting started" page, so it makes sense to include
it directly on the same page.

This also has the advantage of helping the users notice the suggested
pages that will further inform them of the value of Codacy.