clean: Remove reference to admin permissions on source repository
The feature is being changed so that users can select a source
repository where they don't have admin permissions.

The users still need to have admin permissions on the target
repository. However:

- This applies to any change done in the Code patterns page
- To be able to reach the Code patterns page, the users must already
be admins on the repository

As such, I'm no longer mentioning the requirement of the permissions
directly on the "import patterns" feature.