clean: Review Setting a coding standard as default
It's already implicit across the page that Codacy only supports a single
coding standard to be set as the default. So, the extra and explicit
clarification that Codacy only allows one default coding standard makes
sense to be included in the step where users toggle a new coding
standard as default.

Also edits the intro to make it a bit more obvious the distinction
between the "default coding standard configurations" and the "default
Codacy configurations".