clean: Remove tip for setting the coding standard as the default
The ability to set a coding standard as the default is readily visible
on the Codacy UI and is also mentioned on the introduction and on a
dedicated section on the page. I don't think it's necessary to keep it.

Besides this, if we consistently followed this pattern of referencing
related tasks for the same concept at the end of the instructions for
each task, it would become unmanageable to keep these links and it
would probably also confuse the users.