clean: Refactor details on how coding standards work DOCS-556 (#1739)
* wip: Mark all tidbits of information to refactor

* clean: Move limit of 10 coding standards

This is a not so important detail that is mostly relevant if a user is
trying to add a coding standard above the current limit. So, it's better
to include it in the step where users are creating a new coding
standard.

* clean: Review Setting a coding standard as default

It's already implicit across the page that Codacy only supports a single
coding standard to be set as the default. So, the extra and explicit
clarification that Codacy only allows one default coding standard makes
sense to be included in the step where users toggle a new coding
standard as default.

Also edits the intro to make it a bit more obvious the distinction
between the "default coding standard configurations" and the "default
Codacy configurations".

* clean: Remove tip for setting the coding standard as the default

The ability to set a coding standard as the default is readily visible
on the Codacy UI and is also mentioned on the introduction and on a
dedicated section on the page. I don't think it's necessary to keep it.

Besides this, if we consistently followed this pattern of referencing
related tasks for the same concept at the end of the instructions for
each task, it would become unmanageable to keep these links and it
would probably also confuse the users.

* clean: Review admonition for removing coding standards from repositories

Taking the lack of customer feedback into consideration, this admonition
can be a "note" instead of "important".

We could however make it more explicit that the repositories keep the
toggle status of the tools from the coding standard.

* clean: Organize details about assigning coding standards to repositories

Several details are all related to the concept of assigning a coding
standard to repositories, and should therefore be grouped together.

* feat: Add note on restoring configurations to deleting coding standards

While deleting a coding standard, any repositories that were following
that coding standard will recover the previous code pattern
configurations. To make this explicit, I'm including an adapted version
of the note that we already had at the end of the section for editing
a coding standard.

* fix: Fix typo

* clean: Explain default coding standard in concept only

Instead of explaining the value of having a default coding standard
twice, it's better to have a single and full explanation as part of the
concepts.

* clean: Edit and order details in intro

Tries to convey the details in an order that's as close as possible to
a typical chronological order.

* clean: Make sentence more explicit

* clean: Reorder details about applying coding standards to repositories

* clean: Apply suggestions from code review

Co-authored-by: Nicola Klemenc <nicola.klemenc@codacy.com>

---------

Co-authored-by: Nicola Klemenc <nicola.klemenc@codacy.com>